Moved communication tests to tcp subpackage to reflect recent changes
