+++ src/test/java/tests/communication/tcp/TestClientConnection.java
@@ -1,4 +1,4 @@
-package tests.communication;
+package tests.communication.tcp;
 
 import ioopm.mud.communication.rawtcp.server.ClientConnection;
 import org.junit.Before;
